refactor/all

+++ app/presentation/telegram/handlers/admin.py
@@ -1,15 +1,17 @@
-from aiogram import types, Router
+from aiogram import Router, types
 from aiogram.filters import Command
 
 from app.infrastructure.db.repositories import AdminRepository
 from app.presentation.telegram.utils import other
 
+admin_router = Router()
 
-router = Router()
 
-
-@router.message(Command("admin", prefix="!/"))
-async def new_admin(message: types.Message, admin_repo: AdminRepository):
+@admin_router.message(Command("admin", prefix="!/"))
+async def new_admin(message: types.Message, admin_repo: AdminRepository) -> None:
+    if not message.reply_to_message or not message.reply_to_message.from_user:
+        await message.answer("Используйте эту команду в ответ на сообщение пользователя.")
+        return
     target_user = message.reply_to_message.from_user
     if not await admin_repo.is_admin(target_user.id):
         await admin_repo.insert_admin(target_user.id)
@@ -19,8 +21,11 @@ async def new_admin(message: types.Message, admin_repo: AdminRepository):
     await message.delete()
 
 
-@router.message(Command("unadmin", prefix="!/"))
-async def delete_admin(message: types.Message, admin_repo: AdminRepository):
+@admin_router.message(Command("unadmin", prefix="!/"))
+async def delete_admin(message: types.Message, admin_repo: AdminRepository) -> None:
+    if not message.reply_to_message or not message.reply_to_message.from_user:
+        await message.answer("Используйте эту команду в ответ на сообщение пользователя.")
+        return
     target_user = message.reply_to_message.from_user
     if await admin_repo.is_admin(target_user.id):
         await admin_repo.delete_admin(target_user.id)
